Replacing optparse with argparse
Optparse is deprecated, so I did a few modifications
to the code to use this new module.
The only differences are that with argparse the order
of the parameters does not matter.
I place an argument called "script_file" to have the
same initial structure of optparse
(istead of a "-i/--input script_file")

Signed-off-by: Cristian Maureira-Fredes <cmaurei@aei.mpg.de>